// appfront/src/app/api/models/course-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { LessonResponse } from '../models/lesson-response';
export interface CourseResponse {
  categoryName?: string;
  coverImage?: string;
  createdAt?: string;
  description?: string;
  idCourse?: string;
  lessons?: Array<LessonResponse>;
  level?: string;
  price?: number;
  status?: string;
  teacherFullName?: string;
  title?: string;
  totalLessons?: number;
}